-make .po file loading %30 faster thanks to QRegExps caching -fix crash on multiline plurals -mark views' titlebars when they actually display any useful info -source code reorganization
svn path=/trunk/playground/devtools/kaider/; revision=684482

+++ src/catalog/catalogitem.cpp
@@ -184,7 +184,7 @@ bool CatalogItem::isUntranslated() const
 
 bool CatalogItem::isUntranslated(uint form) const
 {
-    if (form<d->_msgstrPlural.size())
+    if ((int)form<d->_msgstrPlural.size())
         return d->_msgstrPlural.at(form).isEmpty();
     else
         return true;
